+++ src/test/java/com/previred/desafio/test/ValorUFHandlerTest.java
@@ -1,4 +1,4 @@
-package com.previred.challenge.test;
+package com.previred.desafio.test;
 
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
@@ -8,7 +8,7 @@ import java.util.List;
 import org.junit.Assert;
 import org.junit.Test;
 
-import com.previred.challenge.ValorUFHandler;
+import com.previred.desafio.ValorUFHandler;
 import com.previred.desafio.tres.uf.vo.Uf;
 
 public class ValorUFHandlerTest {
